chore: update header to 26.51.1

+++ src/mc/world/item/HandSlot.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class HandSlot : int {
+enum class HandSlot : uchar {
     Mainhand = 0,
     Offhand  = 1,
     Count    = 2,
